server exposer as interface with Ingress implementation
