+++ infrastructures/kubernetes/src/main/java/org/eclipse/che/workspace/infrastructure/kubernetes/server/secure/jwtproxy/PassThroughProxySecureServerExposer.java
@@ -17,6 +17,7 @@ import javax.inject.Inject;
 import org.eclipse.che.api.core.model.workspace.runtime.RuntimeIdentity;
 import org.eclipse.che.workspace.infrastructure.kubernetes.environment.KubernetesEnvironment;
 import org.eclipse.che.workspace.infrastructure.kubernetes.server.external.ExternalServerExposer;
+import org.eclipse.che.workspace.infrastructure.kubernetes.server.external.IngressServerExposer;
 import org.eclipse.che.workspace.infrastructure.kubernetes.server.secure.DefaultSecureServerExposer;
 import org.eclipse.che.workspace.infrastructure.kubernetes.server.secure.jwtproxy.factory.PassThroughProxyProvisionerFactory;
 
@@ -25,7 +26,7 @@ import org.eclipse.che.workspace.infrastructure.kubernetes.server.secure.jwtprox
  *
  * <p>To expose secure servers it provisions JwtProxy objects into environment with {@link
  * PassThroughProxyProvisioner}. Then JwtProxy service port is made public accessible by {@link
- * ExternalServerExposer<T>}.
+ * IngressServerExposer <T>}.
  *
  * <p>In this way, requests to exposed secure servers will be routed via JwtProxy pod that is added
  * one per workspace.
