+++ infrastructures/kubernetes/src/test/java/org/eclipse/che/workspace/infrastructure/kubernetes/server/external/MultiHostExternalServiceExposureStrategyTest.java
@@ -46,7 +46,7 @@ public class MultiHostExternalServiceExposureStrategyTest {
   private static final String SERVICE_NAME = SERVER_PREFIX + "12345678" + "-" + MACHINE_NAME;
   private static final String DOMAIN = "che.com";
 
-  private ExternalServerExposer<KubernetesEnvironment> externalServerExposer;
+  private IngressServerExposer<KubernetesEnvironment> externalServerExposer;
   private KubernetesEnvironment kubernetesEnvironment;
 
   @BeforeMethod
@@ -65,7 +65,7 @@ public class MultiHostExternalServiceExposureStrategyTest {
     kubernetesEnvironment =
         KubernetesEnvironment.builder().setPods(ImmutableMap.of("pod", pod)).build();
     externalServerExposer =
-        new ExternalServerExposer<>(
+        new IngressServerExposer<>(
             new MultiHostExternalServiceExposureStrategy(DOMAIN, MULTI_HOST_STRATEGY),
             emptyMap(),
             "%s");
